# demo/openapi/petclinic-rest.yaml
openapi: 3.0.1
info:
  title: PetClinic REST API (TRACE demo excerpt)
  description: >
    Owner/Pet 관련 경로만 발췌한 데모용 OpenAPI 명세.
    원본: spring-petclinic-rest (Apache-2.0), 데모용으로 축약·개변됨.
  version: "1.0.0"
paths:
  /owners:
    get:
      summary: List owners
      operationId: listOwners
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Owner"
    post:
      summary: Create an owner
      operationId: addOwner
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Owner"
      responses:
        "201":
          description: Created
  /owners/{ownerId}:
    get:
      summary: Get an owner by id
      operationId: getOwner
      parameters:
        - name: ownerId
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Owner"
  /owners/{ownerId}/pets:
    get:
      summary: List an owner's pets
      operationId: listPets
      parameters:
        - name: ownerId
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Pet"
  /vets:
    get:
      summary: List veterinarians
      operationId: listVets
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Vet"
  /pets/{petId}/visits:
    post:
      summary: Add a visit for a pet
      operationId: addVisit
      parameters:
        - name: petId
          in: path
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Visit"
      responses:
        "201":
          description: Created
components:
  schemas:
    Owner:
      type: object
      # NOTE(demo): email 속성 없음 (요구사항 REQ-2 신규 정책 미반영 → 충돌 C-3)
      required:
        - firstName
        - lastName
      properties:
        id:
          type: integer
          format: int32
        firstName:
          type: string
          maxLength: 30
        lastName:
          type: string
          maxLength: 30
        address:
          type: string
          maxLength: 255
        city:
          type: string
          maxLength: 80
        telephone:
          type: string
          # NOTE(demo): 요구사항은 20자이나 여기서는 10 (충돌 C-1)
          maxLength: 10
    Pet:
      type: object
      required:
        - name
        - type
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
          maxLength: 30
        birthDate:
          type: string
          format: date
          # NOTE(demo): 미래일자 제약을 스펙에 표기하지 않음 (드리프트 맥락, 충돌 C-2)
        type:
          $ref: "#/components/schemas/PetType"
    PetType:
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
          maxLength: 80
    Vet:
      type: object
      required:
        - firstName
        - lastName
      properties:
        id:
          type: integer
          format: int32
        firstName:
          type: string
          maxLength: 30
        lastName:
          type: string
          maxLength: 30
        # NOTE(demo): specialties 최소 개수 제약을 스펙에 표기하지 않음 (충돌 C-5 맥락)
        specialties:
          type: array
          items:
            type: string
    Visit:
      type: object
      required:
        - date
      properties:
        id:
          type: integer
          format: int32
        date:
          type: string
          format: date
          # NOTE(demo): 과거일자 금지 제약을 스펙에 표기하지 않음 (충돌 C-6 맥락)
        description:
          type: string
          # NOTE(demo): 명세(spec)는 255자이나 여기서는 8192 (충돌 C-4)
          maxLength: 8192